# api/src/main/resources/static/openapi.yml
openapi: 3.0.3

info:
  title: Shopping Cart API
  description: API for managing a user's shopping cart
  version: 1.0.0

servers:
  - url: /api/v1
    description: API base path

tags:
  - name: Cart
    description: Operations related to user shopping cart
  - name: Checkout
    description: Operations related to checkout

paths:
  /users/{userId}/cart:
    get:
      tags: [Cart]
      summary: Get user cart
      description: Retrieve the current cart for a given user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cart retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponseDto'

  /users/{userId}/cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart
      description: Adds an item to the cart. If already present, quantity is increased by 1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddItemRequestDto'
      responses:
        '201':
          description: Item added
        '400':
          description: Invalid input
        '409':
          description: Insufficient stock

    put:
      tags: [Cart]
      summary: Update item quantity
      description: Updates the quantity of an existing product/offer in the user's cart
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequestDto'
      responses:
        '204':
          description: Item quantity updated
        '400':
          description: Invalid input
        '404':
          description: Cart or cart line not found
        '409':
          description: Insufficient stock

  /users/{userId}/cart/items/{productId}/{offerId}:
    delete:
      tags: [Cart]
      summary: Remove item from cart
      description: Removes a product/offer from the cart
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Item removed
        '404':
          description: Cart not found

  /users/{userId}/checkout:
    post:
      tags: [Checkout]
      summary: Checkout cart
      description: Finalize the cart into an order and clear the cart
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Checkout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponseDto'
        '404':
          description: Cart not found
        '409':
          description: Insufficient stock
components:
  schemas:
    AddItemRequestDto:
      type: object
      required:
        - productId
        - offerId
      properties:
        productId:
          type: integer
          format: int64
          example: 10
        offerId:
          type: integer
          format: int64
          example: 5

    UpdateItemRequestDto:
      type: object
      required:
        - productId
        - offerId
        - quantity
      properties:
        productId:
          type: integer
          format: int64
          example: 10
        offerId:
          type: integer
          format: int64
          example: 5
        quantity:
          type: integer
          minimum: 1
          example: 3

    CartResponseDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          example: 42
        lines:
          type: array
          items:
            $ref: '#/components/schemas/CartLineResponseDto'
        totalPrice:
          type: number
          format: double
          example: 129.99

    CartLineResponseDto:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 10
        offerId:
          type: integer
          format: int64
          example: 5
        state:
          type: string
          description: State of the product
          enum: [ NEUF, RECONDITIONNE, OCCASION]
          example: NEUF
        quantity:
          type: integer
          example: 2
        unitPrice:
          type: number
          format: double
          example: 49.99
        lineTotal:
          type: number
          format: double
          example: 99.98